+++ tasks/ax-compass-flow.js
@@ -10,10 +10,10 @@ module.exports = function( grunt ) {
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-   var path = require( 'path' ).posix;
    var debounce = require( 'lodash' ).debounce;
-   var scssHelper = require( './lib/scss-helper' )( grunt, TASK );
    var taskHelper = require( 'grunt-laxar/tasks/lib/task_helpers' )( grunt, TASK );
+   var scssHelper = require( './lib/scss-helper' )( grunt, TASK );
+   var path = require( './lib/path-platform/path' ).posix;
 
    var CONFIG_FILE = path.join( 'work', 'compass-watch-configuration.json' );
    var CSS_MATCHER = /^(.*)[\/\\]css[\/\\](.*)[.]css$/;