+++ tasks/spec/ax-compass-flow.spec.js
@@ -3,10 +3,11 @@
  * Released under the MIT license.
  * http://laxarjs.org/license
  */
-var path = require( 'path' );
+
 var grunt = require( 'grunt' );
 var expect = require( 'expect.js' );
 
+var path = require( '../lib/path-platform/path' ).posix;
 var run = require( './lib/run-elsewhere' );
 
 describe( 'the laxar-compass-flow task', function() {