+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#4](https://github.com/LaxarJS/grunt-laxar-compass/issues/4): allow to rebuild individual artifacts
+    + NEW FEATURE: see ticket for details
 - [#2](https://github.com/LaxarJS/grunt-laxar-compass/issues/2): watch layout SCSS, not just controls/widgets/themes
 - [#1](https://github.com/LaxarJS/grunt-laxar-compass/issues/1): watch all SCSS files for each artifact theme folder, not just the main file