+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#7](https://github.com/LaxarJS/grunt-laxar-compass/issues/7): added documentation for the laxar-compass task
+    + NEW FEATURE: see ticket for details
 - [#5](https://github.com/LaxarJS/grunt-laxar-compass/issues/5): fixed setup instructions for grunt-laxar-compass
 - [#6](https://github.com/LaxarJS/grunt-laxar-compass/issues/6): removed usage of path.posix for compatibility with node v0.10.x
 - [#3](https://github.com/LaxarJS/grunt-laxar-compass/issues/3): the target laxar-compass can now rebuild all artifacts in batch